+++ tomcat/src/main/java/camp/nextstep/controller/AbstractController.java
@@ -1,8 +1,8 @@
 package camp.nextstep.controller;
 
 import camp.nextstep.domain.http.ContentType;
-import camp.nextstep.domain.http.HttpRequest;
-import camp.nextstep.domain.http.HttpResponse;
+import camp.nextstep.domain.http.request.HttpRequest;
+import camp.nextstep.domain.http.response.HttpResponse;
 import camp.nextstep.util.FileUtil;
 
 public abstract class AbstractController implements Controller {
