+++ tomcat/src/main/java/camp/nextstep/controller/LoginController.java
@@ -2,8 +2,8 @@ package camp.nextstep.controller;
 
 import camp.nextstep.db.InMemoryUserRepository;
 import camp.nextstep.domain.http.HttpCookie;
-import camp.nextstep.domain.http.HttpRequest;
-import camp.nextstep.domain.http.HttpResponse;
+import camp.nextstep.domain.http.request.HttpRequest;
+import camp.nextstep.domain.http.response.HttpResponse;
 import camp.nextstep.domain.session.Session;
 import camp.nextstep.domain.session.SessionManager;
 import camp.nextstep.model.User;
