+++ tomcat/src/main/java/camp/nextstep/controller/RegisterController.java
@@ -1,8 +1,8 @@
 package camp.nextstep.controller;
 
 import camp.nextstep.db.InMemoryUserRepository;
-import camp.nextstep.domain.http.HttpRequest;
-import camp.nextstep.domain.http.HttpResponse;
+import camp.nextstep.domain.http.request.HttpRequest;
+import camp.nextstep.domain.http.response.HttpResponse;
 import camp.nextstep.model.User;
 
 public class RegisterController extends AbstractController {
