+++ tomcat/src/main/java/camp/nextstep/controller/RootController.java
@@ -1,8 +1,8 @@
 package camp.nextstep.controller;
 
 import camp.nextstep.domain.http.ContentType;
-import camp.nextstep.domain.http.HttpRequest;
-import camp.nextstep.domain.http.HttpResponse;
+import camp.nextstep.domain.http.request.HttpRequest;
+import camp.nextstep.domain.http.response.HttpResponse;
 
 public class RootController extends AbstractController {
