+++ tomcat/src/main/java/camp/nextstep/domain/http/request/QueryString.java
@@ -1,4 +1,4 @@
-package camp.nextstep.domain.http;
+package camp.nextstep.domain.http.request;
 
 import java.util.AbstractMap.SimpleEntry;
 import java.util.Arrays;
